+++ src/jigsaw/service/time.service.ts
@@ -1,38 +1,75 @@
 import {Time, Moment, WeekTime} from "./time.types";
 import {CommonUtils} from "../core/utils/common-utils";
+
+/**
+ * 时间粒度值
+ */
 export enum TimeGr {
     second, minute, hour, date, week, month
 }
 
+/**
+ * 周起始日期
+ */
 export enum TimeWeekStart {
     sun, mon, tue, wed, thu, fri, sat
 }
 
-export enum TimeUnit{
+/**
+ * 关于时间宏：
+ *
+ * 它通过一些数字，以及语义化的单词作为单位来表示一个时刻，时间宏是一个普通的字符串。
+ * 例如`"now"`这个宏表示当前时刻，`JigsawTime`将会将其换算为其被初始化的时刻。
+ *
+ * 此外，时间宏还支持加减运算。例如`"now-1d"`的意思是昨天的这个时候，相应的，明天则是`"now+1d"`，上个月是`"now-1M"`等。
+ * 时间宏目前可以支持的时刻是`"now"`，支持的所有单位在这个枚举类型中全部列出了。（注意大小写）
+ *
+ * 时间宏的换算是由Jigsaw自动完成的，如果你需要获得一个时间宏的值，请调用`TimeService.getFormatDate()`方法。
+ *
+ * `TimeService`提供了许多有用的时间换算、格式化工具，当你有需要对时间进行运算时，可以参考它的api说。
+ */
+export enum TimeUnit {
     s,m,h,d,w,M,y
-
 }
 
-export enum TimeFormatters{
+/**
+ * 常用时间格式，`TimeService.format`可以支持任何格式，我们做这个枚举只是他们太常用了，使用这个枚举+IDE提示，你可以少敲很多次键盘。
+ */
+export enum TimeFormatters {
     yyyy_mm_dd_hh_mm_ss,yyyy_mm_dd_hh_mm,yyyy_mm_dd_hh,yyyy_mm_dd,yyyy_mm
 }
 
-
+/**
+ * 提供了日期运算、格式化、转换等一系列常用的方法，当你需要对时间做任何操作时，它都可以帮到你。
+ */
 export class TimeService {
+    /**
+     * 将一个时刻转成符合改粒度的值，例如
+     *
+     * ```
+     * TimeService.convertValue("now", TimeGr.date); // -> 2018/3/23 00:00:00
+     * ```
+     *
+     * `"now"`的时分秒被切去。
+     *
+     * @param {WeekTime} value 待转换的时刻，支持时间宏，请参考这里`TimeUnit`的说明。
+     * @param {TimeGr} gr 目标粒度
+     * @return {string} 符合粒度格式的时刻
+     */
     public static convertValue(value: WeekTime, gr:TimeGr): string {
-        value = TimeService.handleWeekValue(value);
+        value = TimeService._handleWeekValue(value);
         value = TimeService.getFormatDate(<Time>value, gr);
         return <string>value;
     }
 
-    private static handleWeekValue(newValue:WeekTime) : Time{
+    private static _handleWeekValue(newValue:WeekTime) : Time{
         if (newValue && typeof newValue["week"] === 'number') {
             return TimeService.getDateFromYearAndWeek(newValue["year"], newValue["week"])
         }
         return <Time>newValue;
     }
 
-    private static timeFormatterConvert(formatter : TimeFormatters):string{
+    private static _timeFormatterConvert(formatter : TimeFormatters):string{
         switch (formatter){
             case TimeFormatters.yyyy_mm_dd_hh_mm_ss : return "YYYY-MM-DD HH:mm:ss";
             case TimeFormatters.yyyy_mm_dd_hh_mm : return "YYYY-MM-DD HH:mm";
@@ -42,20 +79,26 @@ export class TimeService {
         }
     }
 
-    private static timeFormatMap = new Map([
-        [TimeGr.second, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss)],
-        [TimeGr.minute, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm)],
-        [TimeGr.hour, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh)],
-        [TimeGr.date, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd)],
-        [TimeGr.week, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd)],
-        [TimeGr.month, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm)]
+    private static _timeFormatMap = new Map([
+        [TimeGr.second, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss)],
+        [TimeGr.minute, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm)],
+        [TimeGr.hour, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh)],
+        [TimeGr.date, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd)],
+        [TimeGr.week, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd)],
+        [TimeGr.month, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm)]
     ]);
 
-    public static timeUnitConvert(unit : TimeUnit):string{
+    /**
+     * 时间单位枚举值转对应字符串
+     *
+     * @param {TimeUnit} unit 时间单位枚举值
+     * @return {string} 返回对应的字符串
+     */
+    public static timeUnitConvert(unit : TimeUnit):string {
        return TimeUnit[unit];
     }
 
-    private static timeUnitMap = new Map([
+    private static _timeUnitMap = new Map([
         [TimeUnit.s, 'seconds'],
         [TimeUnit.m, 'minutes'],
         [TimeUnit.h, "hours"],
@@ -65,21 +108,25 @@ export class TimeService {
         [TimeUnit.y, 'years']
     ]);
 
-    private static initMoment() {
+    private static _initMoment() {
         try {
             moment.suppressDeprecationWarnings = 1;
-        }catch(e) {
-
+        } catch(e) {
         }
     }
 
-    private static init = TimeService.initMoment();
+    /**
+     * 仅仅是为了初始化moment对象
+     * @type {any}
+     */
+    private static init = TimeService._initMoment();
 
 
     /**
-     * 是否是时间宏字符串
-     * @param time
-     * @returns {boolean}
+     * 判断给定的时刻值是否是时间宏字符串
+     *
+     * @param {Time} time 给定的时刻值
+     * @returns {boolean} 如果给定的值非字符串，则必然返回false。
      */
     public static isMacro(time:Time): boolean {
         if (typeof time === 'string') {
@@ -93,7 +140,7 @@ export class TimeService {
      * @param timeMacro
      * @returns {any}
      */
-    private static convertBasicMacro(timeMacro:string): Date|string {
+    private static _convertBasicMacro(timeMacro:string): Date|string {
         let date;
         switch (timeMacro) {
             case 'now':
@@ -108,39 +155,42 @@ export class TimeService {
 
     /**
      * 时间加减法
-     * @param date  时间
-     * @param num   数量 负数即为减法
-     * @param unit  单位
+     *
+     * @param {Time} date  时间
+     * @param {string|number} num   数量，为负数即为减法
+     * @param {TimeUnit} unit  单位
      */
     public static addDate(date:Time, num:string|number, unit:TimeUnit): Moment {
-        return moment(date).add(num, TimeService.timeUnitMap.get(unit));
+        return moment(date).add(num, TimeService._timeUnitMap.get(unit));
     }
-
 
     /**
      * 根据粒度获取格式化字符串
-     * @param gr
+     *
+     * @param {TimeGr} gr
      * @returns {string}
      */
     public static getFormatter(gr: TimeGr): string {
-        return TimeService.timeFormatMap.has(gr) ? TimeService.timeFormatMap.get(gr) : 'YYYY-MM-DD';
+        return TimeService._timeFormatMap.has(gr) ? TimeService._timeFormatMap.get(gr) : 'YYYY-MM-DD';
     }
 
 
     /**
      * 时间格式化
-     * @param date
-     * @param formatter
+     *
+     * @param {Time} date
+     * @param {string | TimeFormatters} formatter
      */
-    public static format(date:Time, formatter:string|TimeFormatters): string {
-        if(typeof formatter === "number") formatter = TimeService.timeFormatterConvert(formatter);
+    public static format(date: Time, formatter: string | TimeFormatters): string {
+        if(typeof formatter === "number") formatter = TimeService._timeFormatterConvert(formatter);
         return moment(date).format(formatter);
     }
 
     /**
      * 根据粒度格式化时间
-     * @param date
-     * @param gr
+     *
+     * @param {Time} date
+     * @param {TimeGr} gr
      */
     public static formatWithGr(date:Time, gr:TimeGr): string {
         let format = TimeService.getFormatter(gr);
@@ -149,8 +199,9 @@ export class TimeService {
 
 
     /**
-     * 设置默认周开始
-     * @param weekStart
+     * 设置默认周开始，设置之后会影响之后的所有计算结果
+     *
+     * @param {TimeWeekStart} weekStart
      */
     public static setWeekStart(weekStart: TimeWeekStart = TimeWeekStart.sun): void {
         let locale = moment.locale();
@@ -161,23 +212,45 @@ export class TimeService {
         });
     }
 
-
+    /**
+     * 获取给定时间在当年的周数
+     *
+     * @param {Time} date
+     * @return {number}
+     */
     public static getWeekOfYear(date:Time): number {
         return moment(date).week();
     }
 
+    /**
+     * 获取给定时间的年数值
+     *
+     * @param {Time} date
+     * @return {number}
+     */
     public static getYear(date:Time): number {
         return moment(date).year();
     }
 
+    /**
+     * 获取给定时间的月数值
+     *
+     * @param {Time} date
+     * @return {number}
+     */
     public static getMonth(date:Time): number {
         return moment(date).month() + 1;
     }
 
+    /**
+     * 获取给定时间的天数值
+     *
+     * @param {Time} date
+     * @return {number}
+     */
     public static getDay(date:Time): number {
         return moment(date).date();
     }
-
 
     public static getDateFromYearAndWeek(year: number, week: number): Date {
         return moment().year(year).week(week)
@@ -189,9 +262,10 @@ export class TimeService {
 
     /**
      * 根据字符串获取真实时间
-     * @param time
-     * @param gr
-     * @returns {any}
+     *
+     * @param {Time} time
+     * @param {TimeGr} gr
+     * @returns {Time}
      */
     public static getFormatDate(time: Time, gr?: TimeGr): Time {
         let result = time;
@@ -201,16 +275,16 @@ export class TimeService {
             let timeMacroArr = fullPara.exec(time);
             if (timeMacroArr && CommonUtils.isDefined(timeMacroArr[2])) { //有加减 now-2d
                 result = TimeService.addDate(
-                    TimeService.convertBasicMacro(timeMacroArr[1]),
+                    TimeService._convertBasicMacro(timeMacroArr[1]),
                     "" + timeMacroArr[2] + timeMacroArr[3],
                     TimeUnit[timeMacroArr[4]]);
             } else { //无加减 now
-                result = TimeService.convertBasicMacro(time);
+                result = TimeService._convertBasicMacro(time);
             }
         }
         if (typeof result === 'string') {
-            result = TimeService.format(result, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss));
-            result = new Date(moment(result, TimeService.timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss)));
+            result = TimeService.format(result, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss));
+            result = new Date(moment(result, TimeService._timeFormatterConvert(TimeFormatters.yyyy_mm_dd_hh_mm_ss)));
             result = isNaN(result.valueOf()) ? new Date() : result;
         }
         if (result && (gr || (!gr && gr == 0)) || result instanceof Date || result instanceof moment) {